Move dimensions to its own file
It's referenced in many places and not part of of TestCaseGenerator.

As a single file, it can also be reused in Shutterbug.
