+++ src/Dimensions.js
@@ -1,0 +1,7 @@
+export const BANNER = 'banner';
+export const BROWSER = 'browser';
+export const DEVICE = 'device';
+export const OPERATING_SYSTEM = 'operating_system';
+export const ORIENTATION = 'orientation';
+export const RESOLUTION = 'resolution';
+export const ALLOWED_DIMENSIONS = [BANNER, BROWSER, DEVICE, OPERATING_SYSTEM, ORIENTATION, RESOLUTION];
